Split connectors_provider between back and front components
