+++ front/components/agent_builder/capabilities/knowledge/DataSourceViewTable.tsx
@@ -10,7 +10,7 @@ import {
   findSpaceFromNavigationHistory,
 } from "@app/components/data_source_view/context/utils";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
-import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
+import { CONNECTOR_UI_CONFIGURATIONS } from "@app/lib/connector_providers_ui";
 import {
   getDataSourceNameFromView,
   isRemoteDatabase,
@@ -43,7 +43,7 @@ export function DataSourceViewTable({
       spaceDataSourceViews
         .filter((dsv) => {
           const connectorConfig = dsv.dataSource.connectorProvider
-            ? CONNECTOR_CONFIGURATIONS[dsv.dataSource.connectorProvider]
+            ? CONNECTOR_UI_CONFIGURATIONS[dsv.dataSource.connectorProvider]
             : null;
 
           if (connectorConfig?.isHiddenAsDataSource) {
@@ -63,7 +63,7 @@ export function DataSourceViewTable({
           const provider = dsv.dataSource.connectorProvider;
 
           const connectorProvider = provider
-            ? CONNECTOR_CONFIGURATIONS[provider]
+            ? CONNECTOR_UI_CONFIGURATIONS[provider]
             : null;
 
           const icon = provider
