+++ front/components/spaces/EditSpaceManagedDatasourcesViews.tsx
@@ -27,6 +27,7 @@ import SpaceManagedDatasourcesViewsModal from "@app/components/spaces/SpaceManag
 import { useAwaitableDialog } from "@app/hooks/useAwaitableDialog";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
+import { CONNECTOR_UI_CONFIGURATIONS } from "@app/lib/connector_providers_ui";
 import { getDisplayNameForDataSource, isManaged } from "@app/lib/data_sources";
 import { useKillSwitches } from "@app/lib/swr/kill";
 import {
@@ -104,15 +105,15 @@ export function EditSpaceManagedDataSourcesViews({
   const filterSystemSpaceDataSourceViews = useMemo(
     () =>
       systemSpaceDataSourceViews.filter((dsv) => {
-        const connectorConfig = dsv.dataSource.connectorProvider
-          ? CONNECTOR_CONFIGURATIONS[dsv.dataSource.connectorProvider]
+        const connectorUIConfig = dsv.dataSource.connectorProvider
+          ? CONNECTOR_UI_CONFIGURATIONS[dsv.dataSource.connectorProvider]
           : null;
 
         return (
           isManaged(dsv.dataSource) &&
           (!dataSourceView ||
             dsv.dataSource.sId === dataSourceView.dataSource.sId) &&
-          !connectorConfig?.isHiddenAsDataSource
+          !connectorUIConfig?.isHiddenAsDataSource
         );
       }),
     [systemSpaceDataSourceViews, dataSourceView]
